Added DYSms API
